task result

+++ Manager/src/main/kotlin/com/example/manager/controllers/dto/HashTaskStatusResponseDTO.kt
@@ -2,6 +2,6 @@ package com.example.manager.controllers.dto
 
 data class HashTaskStatusResponseDTO (
     val status: String,
-    val progress: Int,
+    val progress: Double,
     val data: List<String>? = null
 )
